Update file structure and remove plugins folder
