+++ client/src/Views/About/About.vue
@@ -37,7 +37,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@use '../assets/styles/variables.scss' as v;
+@use '../../assets/styles/variables.scss' as v;
 
 .about {
   margin: 0 0 100px 0;
