+++ client/src/Views/Home/Carousel.vue
@@ -70,7 +70,7 @@ export default {
       return {
         background:
           "url('" +
-          require(`../assets/images/carousel_${index}.jpg`) +
+          require(`../../assets/images/carousel_${index}.jpg`) +
           "') no-repeat center center",
         backgroundSize: 'cover',
       };
@@ -80,7 +80,7 @@ export default {
 </script>
 
 <style lang="scss">
-@use '../assets/styles/variables.scss' as v;
+@use "../../assets/styles/variables.scss" as v;
 
 .v-sheet {
   display: flex;
