+++ client/src/Views/Reports/FillChart.vue
