--- client/src/plugins/vuetify.js
@@ -1,7 +1,0 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify/lib';
-
-Vue.use(Vuetify);
-
-export default new Vuetify({
-});
